ai-service/requirements.txt
flask
scikit-learn
pandas
numpy
joblib
pytest
waitress
requests
python-dotenv
SpeechRecognition
pydub
